requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
openai
google-api-python-client
google-auth
python-dotenv
python-dateutil
email-validator
pytest
httpx
